+++ lib/dealog_backoffice/accounts/mailer.ex
@@ -1,0 +1,3 @@
+defmodule DealogBackoffice.Accounts.Mailer do
+  use Swoosh.Mailer, otp_app: :dealog_backoffice
+end