+++ lib/dealog_backoffice/accounts/user_notifier.ex
@@ -1,13 +1,18 @@
 defmodule DealogBackoffice.Accounts.UserNotifier do
-  # For simplicity, this module simply logs messages to the terminal.
-  # You should replace it by a proper email or notification tool, such as:
-  #
-  #   * Swoosh - https://hexdocs.pm/swoosh
-  #   * Bamboo - https://hexdocs.pm/bamboo
-  #
-  defp deliver(to, body) do
+  import Swoosh.Email
+  import DealogBackoffice.Gettext
+
+  defp deliver(to, subject, body) do
     require Logger
     Logger.debug(body)
+
+    new()
+    |> to(to)
+    |> from({"DEalog System", "system@dealog.de"})
+    |> subject(subject)
+    |> text_body(body)
+    |> DealogBackoffice.Accounts.Mailer.deliver()
+
     {:ok, %{to: to, body: body}}
   end
 
@@ -15,59 +20,68 @@ defmodule DealogBackoffice.Accounts.UserNotifier do
   Deliver instructions to confirm account.
   """
   def deliver_confirmation_instructions(user, url) do
-    deliver(user.email, """
+    deliver(
+      user.email,
+      gettext("Confirm your DEalog account"),
+      gettext(
+        """
+        Hi %{email},
 
-    ==============================
+        You can confirm your account by visiting the URL below:
 
-    Hi #{user.email},
+        %{url}
 
-    You can confirm your account by visiting the URL below:
-
-    #{url}
-
-    If you didn't create an account with us, please ignore this.
-
-    ==============================
-    """)
+        If you didn't create an account, please ignore this.
+        """,
+        email: user.email,
+        url: url
+      )
+    )
   end
 
   @doc """
   Deliver instructions to reset a user password.
   """
   def deliver_reset_password_instructions(user, url) do
-    deliver(user.email, """
+    deliver(
+      user.email,
+      gettext("Reset your DEalog password"),
+      gettext(
+        """
+        Hi %{user.email},
 
-    ==============================
+        You can reset your password by visiting the URL below:
 
-    Hi #{user.email},
+        %{url}
 
-    You can reset your password by visiting the URL below:
-
-    #{url}
-
-    If you didn't request this change, please ignore this.
-
-    ==============================
-    """)
+        If you didn't request this change, please ignore this.
+        """,
+        email: user.email,
+        url: url
+      )
+    )
   end
 
   @doc """
   Deliver instructions to update a user email.
   """
   def deliver_update_email_instructions(user, url) do
-    deliver(user.email, """
+    deliver(
+      user.email,
+      gettext("Update your DEalog email address"),
+      gettext(
+        """
+        Hi %{user.email},
 
-    ==============================
+        You can change your email address by visiting the URL below:
 
-    Hi #{user.email},
+        %{url}
 
-    You can change your email by visiting the URL below:
-
-    #{url}
-
-    If you didn't request this change, please ignore this.
-
-    ==============================
-    """)
+        If you didn't request this change, please ignore this.
+        """,
+        email: user.email,
+        url: url
+      )
+    )
   end
 end